feat: Add ImageCard and ImageGrid components
The ImageCard component is a new component that renders a card with an image. The ImageGrid component is also a new component that displays a grid of images using the MasonryFlashList component from the @shopify/flash-list library. These components are added to the project to enhance the image display functionality.