```text feat: Add getImageSize function for determining image size
This commit adds a new function, getImageSize, to the common.js helper file. The function takes in the height and width of an image and returns the appropriate size based on whether the image is landscape, portrait, or square. This function will be used in the ImageCard component to set the height of the image dynamically.