feat: Add search functionality to HomeScreen
This commit adds search functionality to the HomeScreen. It includes the implementation of a search bar and the ability to search for images based on user input. The search functionality is triggered when the user types more than 2 characters in the search bar. The search results are displayed in the ImageGrid component. This feature enhances the user experience by allowing them to easily find specific images.