refactor: Update font size in WelcomeScreen title and fix layout in _layout.js
This commit refactors the code to update the font size in the WelcomeScreen title from hp(8) to hp(6) for better visual consistency. Additionally, it fixes the layout in _layout.js by wrapping the content in a GestureHandlerRootView and adding the BottomSheetModalProvider for improved gesture handling and bottom sheet functionality. These changes enhance the user experience and improve the overall appearance of the app.